Replaced rccl_float8 with hip_fp8 and added a local version of hip_fp8 for compatibility with older ROCm versions.
