+++ src/common.cu
@@ -540,7 +540,7 @@ testResult_t startColl(struct threadArgs* args, ncclDataType_t type, ncclRedOp_t
         hip_bfloat16 bf16;
         #endif
         #if defined(RCCL_FLOAT8)
-        rccl_float8 fp8_e4m3; rccl_bfloat8 fp8_e5m2;
+        __hip_fp8_e4m3 fp8_e4m3; __hip_fp8_e5m2 fp8_e5m2;
         #endif
       };
       switch(type) {
@@ -557,8 +557,8 @@ testResult_t startColl(struct threadArgs* args, ncclDataType_t type, ncclRedOp_t
       case ncclBfloat16: bf16 = ncclVerifiablePremulScalar<hip_bfloat16>(rank); break;
       #endif
       #if defined(RCCL_FLOAT8)
-      case ncclFp8E4M3: fp8_e4m3 = ncclVerifiablePremulScalar<rccl_float8>(rank); break;
-      case ncclFp8E5M2: fp8_e5m2 = ncclVerifiablePremulScalar<rccl_bfloat8>(rank); break;
+      case ncclFp8E4M3: fp8_e4m3 = ncclVerifiablePremulScalar<__hip_fp8_e4m3>(rank); break;
+      case ncclFp8E5M2: fp8_e5m2 = ncclVerifiablePremulScalar<__hip_fp8_e5m2>(rank); break;
       #endif
       case ncclNumTypes: break;
       }
